+++ packages/react/src/experimental/OneTable/TableCell/utils/nested.ts
@@ -1,6 +1,11 @@
-export const SPACING_FACTOR = 24
-export const CHEVRON_SIZE = 16
+import { NestedRowProps } from "@/experimental/OneDataCollection/visualizations/collection/Table/components/Row"
+
+export const SPACING_FACTOR = 32
+export const CHEVRON_PARENT_SIZE = 24
+export const CHEVRON_SIZE = 18
 export const LINE_WIDTH = "1px"
+export const PADDING_TOP = 8
+export const BUTTON_HEIGHT = 32
 
 export const getNestedMarginLeft = (depth: number, padding: number = 0) => {
   return `${depth * SPACING_FACTOR + padding}px`
@@ -35,5 +40,22 @@ export const isFirstCellWithNoChildrenAndTableChildren = (
 ) => {
   return (
     !hasChildren && isFirstCellWithTableChildren(firstCell, tableWithChildren)
+  )
+}
+
+export const emptyDetailedCellClassName = (
+  nestedRowProps?: NestedRowProps & { rowWithChildren?: boolean }
+) => {
+  const { nestedVariant, onLoadMoreChildren, rowWithChildren } =
+    nestedRowProps ?? {}
+
+  return (
+    nestedVariant === "detailed" &&
+    !onLoadMoreChildren &&
+    !rowWithChildren &&
+    "has-[span:only-child:empty]:pt-0.5 " +
+      " has-[span:only-child:empty]:before:static" +
+      " has-[span:only-child:empty]:before:opacity-[0.5]" +
+      " has-[span:only-child:empty]:before:content-['N/A']"
   )
 }